requirements.txt
paho-mqtt==1.6.1
nicegui==0.4.2
